maybe fixing table orders?

+++ utils/tables.py
@@ -38,7 +38,7 @@ class ProteinsBase(Reflected, Base):
     __abstract__ = True
 
     gene_id = mapped_column(String(GENE_ID_LENGTH), primary_key=True)
-    gtdb_version = mapped_column(Integer(5), comment='The GTDB gtdb_version from which the data was obtained.')
+    gtdb_version = mapped_column(Integer, comment='The GTDB gtdb_version from which the data was obtained.')
     genome_id = mapped_column(String(GENOME_ID_LENGTH))
     
     seq = mapped_column(String(MAX_GENE_LENGTH), comment='The amino acid sequence.')
@@ -58,7 +58,7 @@ class AnnotationsKeggBase(Reflected, Base):
     __abstract__ = True
  
     annotation_id = mapped_column(Integer, primary_key=True)
-    gtdb_version = mapped_column(Integer(5), comment='The GTDB gtdb_version from which the data was obtained.')
+    gtdb_version = mapped_column(Integer, comment='The GTDB gtdb_version from which the data was obtained.')
     gene_id = mapped_column(String(GENE_ID_LENGTH))
     genome_id = mapped_column(String(GENOME_ID_LENGTH))
 
@@ -72,7 +72,7 @@ class AnnotationsPfamBase(Reflected, Base):
     __abstract__ = True
 
     annotation_id = mapped_column(Integer, primary_key=True)
-    gtdb_version = mapped_column(Integer(5), comment='The GTDB gtdb_version from which the data was obtained.')
+    gtdb_version = mapped_column(Integer, comment='The GTDB gtdb_version from which the data was obtained.')
     gene_id = mapped_column(String(GENE_ID_LENGTH))
     genome_id = mapped_column(String(GENOME_ID_LENGTH))
 
@@ -89,7 +89,7 @@ class MetadataBase(Reflected, Base):
     __abstract__ = True 
 
     genome_id = mapped_column(String(GENOME_ID_LENGTH), primary_key=True)
-    gtdb_version = mapped_column(Integer(5), comment='The GTDB gtdb_version from which the data was obtained.')
+    gtdb_version = mapped_column(Integer, comment='The GTDB gtdb_version from which the data was obtained.')
 
     gtdb_order = mapped_column(String(DEFAULT_STRING_LENGTH))
     gtdb_domain = mapped_column(String(DEFAULT_STRING_LENGTH))
